XF-83 change float to double use zephyr sensor_value_to_double fix tests

+++ src/events/environment/env_sensor_event.h
@@ -11,9 +11,9 @@
 struct env_sensor_event {
 	struct event_header header;
 
-	float temp;
-	float press;
-	float humidity;
+	double temp;
+	double press;
+	double humidity;
 };
 
 struct request_env_sensor_event {
